requirements.txt
fastapi
supabase
uvicorn
python-dotenv